rename Skelanim.hpp to lowercase
