+++ plugins/guacamole-skelanim/include/gua/skelanim.hpp
