+++ plugins/guacamole-skelanim/include/gua/skelanim/platform.hpp
